Update for new PMVHaven site structure
A few changes were made in the recent PMVHaven overhaul that broke the extractor. Videos are now loaded as m3u8, and the direct link to the original video has been obfuscated. This version of the video extractor looks for the direct video URL which is loaded in a javascript segment on the video page. This got complicated because videos uploaded *before* the update are still at `storage.pmvhaven.com`, but newer videos are now at `videos.pmvhaven.com`

User profiles no longer use `/api/users/<username>` and instead use the user ID. The UserIE extracts the user ID from the page then makes the request to `/api/videos?uploader=<id>` for the full list of videos.